Ajustes na questão 2-mOs pacotes de fato são limitados pelos bits
Mudança na função de cálculo de altura da AB

Ajustes nos tipos de dados das funções de print de arquivo

+++ include/binary_tree.h
@@ -49,6 +49,6 @@ int count_nodes(BinTree root);
 
 //Altura da Arvore
 
-uint8_t height_binTree(BinTree root);
+int height_binTree(BinTree root);
 
 #endif // _BINARY_TREE_H
